+++ Tests/Editor/PlayAssetDeliveryBuildWindowsTests.cs
@@ -26,7 +26,7 @@ internal class PlayAssetDeliveryBuildWindowsTests : PlayAssetDeliveryBuildTestsB
     {
         EditorUserBuildSettings.standaloneBuildSubtarget = StandaloneBuildSubtarget.Player;
         EditorUserBuildSettings.SwitchActiveBuildTarget(BuildTargetGroup.Standalone, BuildTarget.StandaloneWindows64);
-        PlayerSettings.SetScriptingBackend(NamedBuildTarget.Standalone, ScriptingImplementation.IL2CPP);
+        PlayerSettings.SetScriptingBackend(NamedBuildTarget.Standalone, ScriptingImplementation.Mono2x);
     }
 
     [TearDown]