+++ Tests/Runtime/AssemblyInfo.cs
@@ -3,5 +3,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: AssemblyCompany("Unity Technologies")]
 
-[assembly: InternalsVisibleTo("Unity.Addressables.Android.Tests")]
 [assembly: InternalsVisibleTo("Unity.Addressables.Android.Editor.Tests")]