+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2025-02-12
+
+* Loading Addressable assets synchronously can now generate an exception. For more information, refer to the documentation.
+* Updated `com.unity.addressables` dependency up to v2.3.16.
+
 ## [1.0.4] - 2024-08-27
 
 Fixed problem with building Addressables content when the empty folder is added to the addressable group.
@@ -22,4 +27,4 @@ Full compatibility with *com.unity.addressables 2.0.3*.
 
 ## [1.0.0] - 2023-09-18
 
-This is the first release of *Unity Package com.unity.addresaables.android*.
+This is the first release of *Unity Package com.unity.addressables.android*.